+++ tests/test-overlay.js
@@ -37,7 +37,8 @@ function kes(imza) {
 var kod = animBlok + "\n" + kes("function assCapa(") + "\n" +
   kes("function assRenk(") + "\n" + kes("function buildAss(") + "\n" +
   kes("function assTc(") + "\n" + kes("function assMetin(") + "\n" +
-  kes("function assKaraokeSatirlari(") + "\n" +
+  kes("function assKaraokeSatirlari(") + "\n" + kes("function stilRecetesi(") + "\n" +
+  kes("function yogunlukCarpani(") + "\n" + kes("function vurguKelimesi(") + "\n" +
   "; return { buildAss: buildAss };";
 
 var gecti = 0, kaldi = 0;